+++ app/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.ui.ApplicantListActivity"></activity>
+        <activity android:name=".View.ui.ApplicantListActivity"></activity>
 
         <meta-data
             android:name="com.kakao.sdk.AppKey"
             android:value="@string/kakao_app_key" />
 
-        <activity android:name=".view.ui.LoginActivity2" />
-        <activity android:name=".view.ui.JoinRoomActivity">
+        <activity android:name=".View.ui.LoginActivity2" />
+        <activity android:name=".View.ui.JoinRoomActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
 
@@ -31,9 +31,9 @@
                     android:scheme="@string/kakao_scheme" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.ui.CreateRoomActivity"></activity>
-        <activity android:name=".view.ui.MainActivity" />
-        <activity android:name=".view.ui.LoginActivity">
+        <activity android:name=".View.ui.CreateRoomActivity"></activity>
+        <activity android:name=".View.ui.MainActivity" />
+        <activity android:name=".View.ui.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
