+++ app/src/main/java/com/example/mnt_android/View/ui/CreateRoomFragment2.kt
@@ -1,17 +1,13 @@
-package com.example.mnt_android.view.ui
+package com.example.mnt_android.View.ui
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProviders
 import com.example.mnt_android.R
-import com.example.mnt_android.databinding.ActivityCreateroomBinding
 import com.example.mnt_android.databinding.FragmentCreateroom2Binding
 import com.example.mnt_android.viewmodel.CreateRoomViewModel
