+++ app/src/main/java/com/example/mnt_android/View/ui/JoinRoomActivity.kt
@@ -1,6 +1,5 @@
-package com.example.mnt_android.view.ui
+package com.example.mnt_android.View.ui
 
-import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
@@ -9,10 +8,7 @@ import androidx.fragment.app.FragmentTransaction
 import androidx.lifecycle.ViewModelProviders
 import com.example.mnt_android.R
 import com.example.mnt_android.viewmodel.JoinRoomViewModel
-import com.example.mnt_android.view.ui.CreateRoomFragment
-import com.example.mnt_android.view.ui.CreateRoomFragment2
 import com.example.mnt_android.viewmodel.BackPressViewModel
-import com.example.mnt_android.viewmodel.CreateRoomViewModel
 
 class JoinRoomActivity : AppCompatActivity()
 {
