+++ app/src/main/java/com/example/mnt_android/View/ui/JoinRoomFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mnt_android.view.ui
+package com.example.mnt_android.View.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,19 +7,19 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.example.mnt_android.R
-import com.example.mnt_android.databinding.FragmentJoinroom2Binding
 import com.example.mnt_android.viewmodel.JoinRoomViewModel
+import com.example.mnt_android.databinding.FragmentJoinroom1Binding
 
-class JoinRoomFragment2 : Fragment()
+class JoinRoomFragment : Fragment()
 {
     lateinit var joinRoomViewModel: JoinRoomViewModel
-    lateinit var binding: FragmentJoinroom2Binding
+    lateinit var binding: FragmentJoinroom1Binding
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_joinroom2,container,false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_joinroom1,container,false)
 
         joinRoomViewModel=(activity as JoinRoomActivity).joinRoomViewModel
