+++ app/src/main/java/com/example/mnt_android/View/ui/MainActivity.kt
@@ -1,13 +1,11 @@
-package com.example.mnt_android.view.ui
+package com.example.mnt_android.View.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProviders
 import com.example.mnt_android.R
-import com.example.mnt_android.databinding.ActivityLogin2Binding
 import com.example.mnt_android.databinding.ActivityMainBinding
 import com.example.mnt_android.viewmodel.BackPressViewModel
 import com.example.mnt_android.viewmodel.LoginViewModel
