+++ app/src/main/java/com/example/mnt_android/View/ui/SplashActivity.kt
@@ -1,0 +1,2 @@
+package com.example.mnt_android.View.ui
+
