+++ app/src/main/java/com/example/mnt_android/View/viewholder/ApplicantListViewHolder.kt
@@ -1,10 +1,9 @@
-package com.example.mnt_android.view.viewholder
+package com.example.mnt_android.View.viewholder
 
 import android.view.View
-import android.widget.Toast
 import androidx.fragment.app.FragmentActivity
 import com.example.mnt_android.base.BaseViewHolder
-import com.example.mnt_android.view.dialog.CustomDialog
+import com.example.mnt_android.View.dialog.CustomDialog
 import kotlinx.android.synthetic.main.item_applicant.view.*
 
 class ApplicantListViewHolder(view: View) : BaseViewHolder(view) {
