+++ app/src/main/java/com/example/mnt_android/ViewModel/CreateRoomViewModel.kt
@@ -1,17 +1,8 @@
 package com.example.mnt_android.viewmodel
 
 import android.util.Log
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.mnt_android.service.model.RoomInfo
-import com.example.mnt_android.view.ui.CreateRoomActivity
-import com.kakao.kakaolink.v2.KakaoLinkResponse
-import com.kakao.kakaolink.v2.KakaoLinkService
-import com.kakao.message.template.LinkObject
-import com.kakao.message.template.TextTemplate
-import com.kakao.network.ErrorResult
-import com.kakao.network.callback.ResponseCallback
 import kotlin.random.Random
 
 class CreateRoomViewModel : ViewModel()
