--- app/src/main/java/com/example/mnt_android/view/ui/SplashActivity.kt
@@ -1,2 +1,0 @@
-package com.example.mnt_android.view.ui
-
