+++ app/src/main/res/layout/activity_applicant_list.xml
@@ -9,7 +9,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.ui.ApplicantListActivity">
+        tools:context=".View.ui.ApplicantListActivity">
         <Button
             android:id="@+id/close_btn"
             android:layout_width="50dp"
