+++ app/src/main/res/layout/activity_createroom.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="createRoomActivity"
-            type="com.example.mnt_android.view.ui.CreateRoomActivity" />
+            type="com.example.mnt_android.View.ui.CreateRoomActivity" />
         <variable
             name="createRoomViewModel"
             type="com.example.mnt_android.viewmodel.CreateRoomViewModel" />
