+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ui.LoginActivity"
+    tools:context=".View.ui.LoginActivity"
     >
 
     <com.kakao.usermgmt.LoginButton
